% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{background_sample}
\alias{background_sample}
\title{Generate pseudo-absence points}
\usage{
background_sample(region, N, ...)
}
\arguments{
\item{region}{The region in which to generate pseudo-absence points. Must be
an object with an \code{\link[sf:st_sample]{sf::st_sample()}} method, or an object
coercible with \link[sf:st_bbox]{sf::st_bbox} (for much faster sampling from
a rectangular bounding box).}

\item{N}{integer. Number of pseudo-absence points to generate.}

\item{...}{Name-value pairs of expressions to be added as fixed attributes
of the pseudo-absences. See \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\value{
An \code{sf} object containing \code{N} random points within \code{region}, with attributes
specified in \code{...}.
}
\description{
Generate pseudo-absence points
}
