% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BadiaPaleoFloraENM-package.R
\docType{package}
\name{BadiaPaleoFloraENM-package}
\alias{BadiaPaleoFloraENM}
\alias{BadiaPaleoFloraENM-package}
\title{BadiaPaleoFloraENM: Ecological Niche Modelling of Paleofauna in the *Badia*}
\description{
Compendium of data and code for our paper ....
}
\author{
\strong{Maintainer}: Joe Roe \email{joeroe@hey.com}

}
\keyword{internal}
