% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{bbox_wkt}
\alias{bbox_wkt}
\title{Well-Known Text (WKT) representation of a bounding box}
\usage{
bbox_wkt(bbox)
}
\arguments{
\item{bbox}{\link[sf:st_bbox]{sf::st_bbox} object}
}
\description{
Well-Known Text (WKT) representation of a bounding box
}
