% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{buffer_bbox}
\alias{buffer_bbox}
\title{Buffer a bounding box}
\usage{
buffer_bbox(bbox, add)
}
\description{
Buffer a bounding box
}
\examples{
w_asia <- st_bbox(c(xmin = 30, xmax = 50, ymin = 25, ymax = 40))
buffer_bbox(w_asia, 10)
}
