% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{extract_points}
\alias{extract_points}
\title{Extract point values from a stars raster}
\usage{
extract_points(raster, points)
}
\arguments{
\item{raster}{\code{stars} object}

\item{points}{\code{sf} object with point geometry}
}
\value{
A data frame with the same number of rows as \code{points}.
}
\description{
Analogous to raster::extract() / terra::extract(), but for stars format rasters.
}
