% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_gbif}
\alias{fetch_gbif}
\title{Fetch GBIF data}
\usage{
fetch_gbif(genus, species, ext, cache_path, redownload = FALSE)
}
\arguments{
\item{genus}{Character. Genus name.}

\item{species}{Character. Species name. Use '*' for fuzzy matching.}

\item{ext}{\code{extent} object. Geographic extent of records to return.}

\item{cache_path}{Character. Path to cache data. Downloaded data is cached as
a CSV with an automatically generated filename.}

\item{redownload}{Logical. If \code{TRUE}, cached data will be ignored. Default: \code{FALSE}.}
}
\value{
A \code{tibble} of downloaded GBIF records.
}
\description{
A simplified interface to \code{\link[dismo:gbif]{dismo::gbif()}} with caching.
}
\details{
The return value is always read from the cached file to ensure consistency.
See \code{\link[dismo:gbif]{dismo::gbif()}} for more details.
}
