% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fetch_paleoclim}
\alias{fetch_paleoclim}
\title{Fetch PaleoClim data}
\usage{
fetch_paleoclim(period, resolution, ext = NULL, cache_file, redownload = FALSE)
}
\arguments{
\item{period}{Character. Time period to fetch.}

\item{resolution}{Character. Resolution to fetch.}

\item{ext}{\code{Extent} object. Region to fetch.}

\item{cache_file}{Character. Path to cache data in .grd format.}

\item{redownload}{Logical. If \code{TRUE}, cached data will be ignored. Default: \code{FALSE}.}
}
\value{
\code{RasterStack} object.
}
\description{
An interface to \code{\link[rpaleoclim:paleoclim]{rpaleoclim::paleoclim()}} with caching.
}
\details{
The raster is always read or re-read from the cached file to ensure consistency.
See \code{\link[rpaleoclim:paleoclim]{rpaleoclim::paleoclim()}} for more details.
}
