% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{gbif_accepted_name}
\alias{gbif_accepted_name}
\title{Accepted name of a taxon}
\usage{
gbif_accepted_name(name)
}
\arguments{
\item{name}{Character. Name of a taxon.}
}
\value{
If the GBIF Backbone Taxonomy considers \code{name} a synonym, the canonical name
of the accepted usage. Otherwise, \code{name}.
}
\description{
Resolves the accepted name of a given taxon according to the GBIF Backbone
Taxonomy.
}
