% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{gbif_drop_imprecise_coords}
\alias{gbif_drop_imprecise_coords}
\alias{gbif_drop_duplicate_coords}
\alias{gbif_drop_fossils}
\alias{gbif_standardise_names}
\title{Clean GBIF data}
\usage{
gbif_drop_imprecise_coords(data, threshold = 5000)

gbif_drop_duplicate_coords(data)

gbif_drop_fossils(data)

gbif_standardise_names(data, shorten = TRUE)
}
\arguments{
\item{data}{Table of occurrence data retrieved with \code{\link[rgbif:occ_data]{rgbif::occ_data()}}}

\item{threshold}{For \code{gbif_drop_imprecise_coords()}, the minimum acceptable
coordinate uncertainty (in meters).}

\item{shorten}{If \code{TRUE}, \code{gbif_standardise_names()} shorten long column
names in addition to converting them to snake case, e.g. \code{decimalLongitude}
to \code{longitude}.}
}
\value{
Modified version of \code{data}.
}
\description{
A collection of functions for cleaning occurrence data from GBIF.
}
