% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{gbif_to_sf}
\alias{gbif_to_sf}
\title{Convert GBIF occurrence data to sf}
\usage{
gbif_to_sf(data)
}
\arguments{
\item{data}{Table of occurrence data retrieved with \code{\link[rgbif:occ_data]{rgbif::occ_data()}}}
}
\value{
An \link[sf:sf]{sf::sf} object.
}
\description{
Converts occurrence data retrieved with \code{\link[rgbif:occ_data]{rgbif::occ_data()}} to a
\link[sf:sf]{sf::sf}-format simple features object with point geometry.
}
\details{
Assumes that \code{data} has been pre-processed with \code{\link[=gbif_standardise_names]{gbif_standardise_names()}}
and does not include rows with missing coordinates.
}
