% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm.R
\name{generate_psabsences}
\alias{generate_psabsences}
\title{Generate pseudo-absence points}
\usage{
generate_psabsences(region, ..., N = 10000)
}
\arguments{
\item{region}{object of class \code{sf} or \code{sfc}. The region in which to generate
pseudo-absence points.}

\item{...}{Name-value pairs of expressions to be added as fixed attributes
of the pseudo-absences. See \code{\link[dplyr:mutate]{dplyr::mutate()}}.}

\item{N}{integer. Number of pseudo-absence points to generate.}
}
\description{
Generate pseudo-absence points
}
