% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_bioclim}
\alias{get_bioclim}
\title{Regional Bioclimatic Variables}
\usage{
get_bioclim(extent, resolution = 0.5, cache_path = getwd(),
  quiet = FALSE)
}
\arguments{
\item{extent}{A \link[raster:Extent]{raster::Extent} object specifying the region of interest,
or an object (e.g. matrix, Spatial*) that can be passed to
\code{\link[raster:extent]{raster::extent()}}}

\item{resolution}{Numeric. Must be 0.5, 2.5, 5 or 10 (minutes of arc). Default: 0.5.}

\item{cache_path}{Character. Path to directory to cache downloaded rasters.}

\item{quiet}{Logical. If \code{TRUE}, suppresses information messages. Default: \code{FALSE}.}
}
\description{
Downloads WorldClim data and derives bioclimatic variables for an arbitrary
region (or extent).
}
