% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hack.R
\name{hack}
\alias{hack}
\title{Workaround for broken slice.stars_proxy:
https://github.com/r-spatial/stars/issues/751}
\usage{
hack(.data, ...)
}
\description{
Workaround for broken slice.stars_proxy:
https://github.com/r-spatial/stars/issues/751
}
