% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.R
\name{logit_to_p}
\alias{logit_to_p}
\title{Convert log-odds to probability}
\usage{
logit_to_p(x)
}
\arguments{
\item{x}{Numeric. Vector of log-odds, e.g. output of a logistic regression.}
}
\value{
Vector of normalised probabilities.
}
\description{
Convert log-odds to probability
}
