% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_if_needed}
\alias{merge_if_needed}
\title{Merge Raster* objects (if necessary)}
\usage{
merge_if_needed(rasters)
}
\arguments{
\item{rasters}{List of \verb{Raster*} objects.}
}
\description{
Utility function for merging a list of Raster* objects. If passed a list of
length 1, simply extracts and returns the first element.
}
