% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{raw_data}
\alias{raw_data}
\alias{derived_data}
\title{Path helpers}
\usage{
raw_data(...)

derived_data(...)
}
\arguments{
\item{...}{Additional path components passed to \code{\link[here:here]{here::here()}}}
}
\description{
Helper functions to construct paths relative to the root of the research
compendium.
}
\examples{
raw_data("ne")
derived_data("gbif")
}
