% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{raster_burn}
\alias{raster_burn}
\title{Burn points into a raster}
\usage{
raster_burn(points, raster)
}
\arguments{
\item{points}{\code{sf}.}

\item{raster}{}
}
\description{
Burn points into a raster
}
